+++ src/commonMain/kotlin/King.kt
@@ -1,9 +1,9 @@
 class King(override val color: PieceColor) : Piece(PieceType.KING, color, if (color == PieceColor.WHITE) 'K' else 'k') {
     override fun canMove(move: PotentialMove): Boolean {
-        if (move.distanceX > 1.toUByte() || move.distanceY > 1.toUByte()) {
+        if (move.distanceX > 1.toUInt() || move.distanceY > 1.toUInt()) {
             return false
         }
-        return move.deltaY == 0.toByte() || move.deltaX == 0.toByte() || move.distanceY == move.distanceX
+        return move.deltaY == 0 || move.deltaX == 0 || move.distanceY == move.distanceX
     }
 
     override val trajectories: List<MoveTrajectory> =
