+++ src/commonMain/kotlin/Knight.kt
@@ -1,7 +1,7 @@
 class Knight(override val color: PieceColor) :
     Piece(PieceType.KNIGHT, color, if (color == PieceColor.WHITE) 'N' else 'n') {
     override fun canMove(move: PotentialMove): Boolean {
-        return move.distanceX == 2.toUByte() && move.distanceY == 1.toUByte() || move.distanceX == 1.toUByte() && move.distanceY == 2.toUByte()
+        return move.distanceX == 2.toUInt() && move.distanceY == 1.toUInt() || move.distanceX == 1.toUInt() && move.distanceY == 2.toUInt()
     }
 
     override val trajectories: List<MoveTrajectory> = listOf(MoveTrajectory.KNIGHT)
