+++ src/commonMain/kotlin/Rook.kt
@@ -2,7 +2,7 @@ class Rook(override val color: PieceColor) :
     Piece(PieceType.ROOK, color, if (color === PieceColor.WHITE) 'R' else 'r') {
 
     override fun canMove(move: PotentialMove): Boolean {
-        return move.deltaY == 0.toByte() || move.deltaX == 0.toByte()
+        return move.deltaY == 0 || move.deltaX == 0
     }
 
     override val trajectories: List<MoveTrajectory> = listOf(MoveTrajectory.VERTICAL, MoveTrajectory.HORIZONTAL)
